requirements.txt
# Core dependencies for a typical Python bot project
python-dotenv
requests
aiohttp
discord.py